+++ src/Beef.Events/WebJobs/WebJobsBuilderExtensions.cs
@@ -36,14 +36,15 @@ namespace Beef.Events.WebJobs
         }
 
         /// <summary>
-        /// Gets (and builds on first access) the <see cref="IConfiguration"/>. Builds the configuration probing; will probe in the following order: 1) Azure Key Vault (see https://docs.microsoft.com/en-us/aspnet/core/security/key-vault-configuration) 
-        /// or User Secrets where hosting environment is development (see https://docs.microsoft.com/en-us/aspnet/core/security/app-secrets), 2) environment variable (see <paramref name="environmentVariablePrefix"/>),
-        /// 3) appsettings.{environment}.json, 4) appsettings.json, 5) {<paramref name="embeddedFilePrefix"/>}.{environment}.json (embedded resource), and 6) {<paramref name="embeddedFilePrefix"/>}.json (embedded resource).
+        /// Gets (and builds on first access) the <see cref="IConfiguration"/>. 
+        /// Builds the configuration probing; will probe in the following order: 1) Azure Key Vault (see https://docs.microsoft.com/en-us/aspnet/core/security/key-vault-configuration),
+        /// 2) User Secrets where hosting environment is development (see https://docs.microsoft.com/en-us/aspnet/core/security/app-secrets), 3) environment variable (see <paramref name="environmentVariablePrefix"/>),
+        /// 4) appsettings.{environment}.json, 5) appsettings.json, 6) webapisettings.{environment}.json (embedded resource), and 7) webapisettings.json (embedded resource). 
         /// </summary>
         /// <typeparam name="TStartup">The function startup <see cref="Type"/>.</typeparam>
         /// <param name="builder">The <see cref="IWebJobsBuilder"/>.</param>
         /// <param name="embeddedFilePrefix">The embedded resource name prefix (defaults to "webjobs").</param>
-        /// <param name="environmentVariablePrefix">The environment variables prefix.</param>
+        /// <param name="environmentVariablePrefix">The prefix that the environment variables must start with (will automatically add a trailing underscore where not supplied).</param>
         /// <param name="configurationBuilder">Action to enable further configuration sources to be added.</param>
         /// <returns>The <see cref="IConfiguration"/> instance.</returns>
         public static IConfiguration GetConfiguration<TStartup>(this IWebJobsBuilder builder, string embeddedFilePrefix = "webjobs",
@@ -63,27 +64,28 @@ namespace Beef.Events.WebJobs
                 cb.AddJsonFile(new EmbeddedFileProvider(typeof(TStartup).Assembly), $"{embeddedFilePrefix}.json", true, false)
                     .AddJsonFile(new EmbeddedFileProvider(typeof(TStartup).Assembly), $"{embeddedFilePrefix}.{hostingEnvironment}.json", true, false)
                     .AddJsonFile("appsettings.json", true, true)
-                    .AddJsonFile($"appsettings.{hostingEnvironment}.json", true, true)
-                    .AddEnvironmentVariables(environmentVariablePrefix);
+                    .AddJsonFile($"appsettings.{hostingEnvironment}.json", true, true);
+
+                if (string.IsNullOrEmpty(environmentVariablePrefix))
+                    cb.AddEnvironmentVariables();
+                else 
+                    cb.AddEnvironmentVariables(environmentVariablePrefix.EndsWith("_", StringComparison.InvariantCulture) ? environmentVariablePrefix : environmentVariablePrefix + "_");
 
                 configurationBuilder?.Invoke(cb);
 
                 // Build as new and replace existing.
                 var config = cb.Build();
-                if (string.Compare(hostingEnvironment, "development", StringComparison.InvariantCultureIgnoreCase) == 0)
+                if (string.Compare(hostingEnvironment, "development", StringComparison.InvariantCultureIgnoreCase) == 0 && config.GetValue<bool>("UseUserSecrets"))
+                    cb.AddUserSecrets<TStartup>();
+
+                var kvn = config["KeyVaultName"];
+                if (!string.IsNullOrEmpty(kvn))
                 {
-                    if (config.GetValue<bool>("UseUserSecrets"))
-                        cb.AddUserSecrets<TStartup>();
-                }
-                else
-                {
-                    var kvn = config["KeyVaultName"];
-                    if (!string.IsNullOrEmpty(kvn))
-                    {
-                        var astp = new AzureServiceTokenProvider();
-                        using var kvc = new KeyVaultClient(new KeyVaultClient.AuthenticationCallback(astp.KeyVaultTokenCallback));
-                        cb.AddAzureKeyVault($"https://{kvn}.vault.azure.net/", kvc, new DefaultKeyVaultSecretManager());
-                    }
+                    var astp = new AzureServiceTokenProvider();
+#pragma warning disable CA2000 // Dispose objects before losing scope; this object MUST NOT be disposed or will result in further error - only a single instance so is OK.
+                    var kvc = new KeyVaultClient(new KeyVaultClient.AuthenticationCallback(astp.KeyVaultTokenCallback));
+                    cb.AddAzureKeyVault($"https://{kvn}.vault.azure.net/", kvc, new DefaultKeyVaultSecretManager());
+#pragma warning restore CA2000
                 }
 
                 // Build again and replace existing.